[ICU-4189] Fix previous change for Solaris with Workshpo 4.2. X-Commit-URL: https://ssl.icu-project.org/trac/changeset/17158 Former-commit-id: 4ef46383f32c187aa7815a92ca963600ec0a8dcf

+++ icu4c/source/i18n/translit.cpp
@@ -790,7 +790,7 @@ UnicodeString& U_EXPORT2 Transliterator::getDisplayName(const UnicodeString& id,
                 uprv_strcpy(key, RB_SCRIPT_DISPLAY_NAME_PREFIX);
                 args[j].getString(s);
                 if (uprv_isInvariantUString(s.getBuffer(), s.length())) {
-                    s.extract(0, sizeof(key)-length-1, key+length, sizeof(key)-length-1, US_INV);
+                    s.extract(0, sizeof(key)-length-1, key+length, (int32_t)sizeof(key)-length-1, US_INV);
 
                     resString = bundle.getStringEx(key, status);
